:test_tube: Fix test failures
There were so many things i was doing wrong in prepping
the environment for tests and also on the side of
creating the database models.

The errors are now fixed, and I will look into creating
a CI action on github to run the test on each push and
pull request to main.