:test_tube: Fix insertion of users logic
Previous state of the code will allow insertion of incase
a username is not associated with the email passed. Meaning
there can be duplicate usernames and emails assocaited to
different accounts. Which will actually cause errors on the
database during insertion.

To fix this, I did a separate queries to find a user by the
username passed and another by the email passed. If any of
them finds a user, the function for checking if a user exists
will return True. If none of them finds a user then False.

Forgot to do the check if the username query count is = to 1
and that of the password query. This is an ammend.