:ambulance: Fix user insertion endpoint
The user insertion endpoint was throwing errors during insertion,
because I didn't import the models whiles creating the session
(has something to do with the declarative base thing of sqlalchemy)

Also I created another pydantic model as the return type of the data
the user will be getting. I did this in order to remove the password
when returning the same data and also for it to reflect in the api
docs generated for us. Using the exlude option of the post method
won't change it on the docs directly.