:adhesive_bandage: Use custom response
From start I was getting errors on serializing my models to
jSON so I serialize them normally before passing it through
the JSON response. This caused it to add extra slashes to the
strings (as in escaping the ""). To prevent this am now using
the Response object to create my custom response passing in
the media type application/json.